# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Grupo5 Usuarios API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the Grupo5 Usuarios service."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "List users",
        "responses": {
          "200": {
            "description": "Users list"
          }
        }
      },
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre_completo",
                  "email",
                  "password",
                  "nacionalidad"
                ],
                "properties": {
                  "nombre_completo": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "rol": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "interno",
                      "usuario"
                    ]
                  },
                  "nacionalidad": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    }
  }
}